Clean api surface

+++ src/synth/reverb.rs
@@ -9,7 +9,7 @@ impl Synth {
     /**
     Set the parameters for the built-in reverb unit
      */
-    pub fn set_reverb_params(&mut self, roomsize: f64, damp: f64, width: f64, level: f64) {
+    pub fn set_reverb_params(&mut self, roomsize: f32, damp: f32, width: f32, level: f32) {
         self.handle.set_reverb_params(roomsize, damp, width, level);
     }
 
@@ -30,28 +30,28 @@ impl Synth {
     /**
     Query the current reverb room size
      */
-    pub fn get_reverb_roomsize(&self) -> f64 {
+    pub fn get_reverb_roomsize(&self) -> f32 {
         self.handle.get_reverb_roomsize()
     }
 
     /**
     Query the current reverb dumping
      */
-    pub fn get_reverb_damp(&self) -> f64 {
+    pub fn get_reverb_damp(&self) -> f32 {
         self.handle.get_reverb_damp()
     }
 
     /**
     Query the current reverb level
      */
-    pub fn get_reverb_level(&self) -> f64 {
+    pub fn get_reverb_level(&self) -> f32 {
         self.handle.get_reverb_level()
     }
 
     /**
     Query the current reverb width
      */
-    pub fn get_reverb_width(&self) -> f64 {
+    pub fn get_reverb_width(&self) -> f32 {
         self.handle.get_reverb_width()
     }
